update to bevy 0.19

+++ src/ui/separator_custom.rs
@@ -62,7 +62,7 @@ impl Widget for SeparatorCustom {
             let stroke = ui.visuals().window_stroke;
             let mut painter = ui.painter().clone();
             let wnd_margin = ui.style().spacing.window_margin;
-            let rect = rect.expand2(vec2(wnd_margin.left, wnd_margin.top));
+            let rect = rect.expand2(vec2(wnd_margin.left as f32, wnd_margin.top as f32));
             painter.set_clip_rect(painter.clip_rect().expand(3.0));
             if is_horizontal_line {
                 painter.hline(
